Adding bean validation on entities based on https://deltaspike.apache.org/documentation/bean-validation.html
Currently it does NOT return the actual validation message as it looks. What it does is just says :
Caused by: javax.validation.ConstraintViolationException: Bean Validation constraint(s) violated while executing Automatic Bean Validation on callback event:'prePersist'. Please refer to embedded ConstraintViolations for details.

This is an issue that has to be researched, maybe we use old version of deltaspike ?